Do not import a package of substituted libraries
The behavior of expecting package imports for the dependency reactor should align with the dependency netty and its package imports.

Both dependencies get substituted and pushed through the shade plugin.

The result is a fat jar where netty and reactor get embedded in the package org.neo4j.driver.internal.shaded.

Note
Although this is convenient, the packaging option should be left to the client and not be included in a library. This is even more important if you want to be able to ship the library for different native runtime options in the future.